vqa-framework/requirements.txt
black==22.3.0
boto3==1.20.24
GitPython==3.1.18
h5py==3.6.0
imageio==2.9.0
matplotlib==3.5.0
Pillow==9.0.1
plotly==5.6.0
pytest==7.1.1
sexpdata==0.0.3
scipy==1.7.3
scikit-learn==1.0.2
tqdm==4.63.0
wget==3.2
opencv-python==4.5.5.64
mlflow==1.22.0
optuna==2.10.0
pytorch-lightning==1.6.0
-f https://download.pytorch.org/whl/torch_stable.html 
torch==1.10.1+cu102
-f https://download.pytorch.org/whl/torch_stable.html 
torchaudio==0.10.1+cu102
-f https://download.pytorch.org/whl/torch_stable.html 
torchvision==0.11.2+cu102
